+++ src/main/java/io/choerodon/devops/app/service/impl/DevopsSecretServiceImpl.java
@@ -96,7 +96,7 @@ public class DevopsSecretServiceImpl implements DevopsSecretService {
         V1Secret v1Secret = initV1Secret(devopsSecretE);
 
         // 更新操作如果key-value没有改变
-        if ("update".equals(secretReqDTO.getType())) {
+        if (UPDATE.equals(secretReqDTO.getType())) {
             DevopsSecretE oldSecretE = devopsSecretRepository
                     .selectByEnvIdAndName(secretReqDTO.getEnvId(), secretReqDTO.getName());
             Map<String, String> oldMap = new HashMap<>();
@@ -111,15 +111,13 @@ public class DevopsSecretServiceImpl implements DevopsSecretService {
 
         // 在gitops库处理secret文件
         operateEnvGitLabFile(TypeUtil.objToInteger(devopsEnvironmentE.getGitlabEnvProjectId()), v1Secret, devopsSecretE,
-                devopsEnvCommandE, "create".equals(secretReqDTO.getType()), userAttrE);
+                devopsEnvCommandE, CREATE.equals(secretReqDTO.getType()), userAttrE);
 
         return ConvertHelper.convert(devopsSecretRepository.queryBySecretId(devopsSecretE.getId()), SecretRepDTO.class);
     }
 
     private DevopsSecretE handleSecret(SecretReqDTO secretReqDTO) {
-        DevopsEnvironmentE devopsEnvironmentE = devopsEnvironmentRepository.queryById(secretReqDTO.getEnvId());
-
-        if ("create".equals(secretReqDTO.getType())) {
+        if (CREATE.equals(secretReqDTO.getType())) {
             // 校验secret名字合法性和环境下唯一性
             DevopsSecretValidator.checkName(secretReqDTO.getName());
             devopsSecretRepository.checkName(secretReqDTO.getName(), secretReqDTO.getEnvId());
@@ -136,7 +134,7 @@ public class DevopsSecretServiceImpl implements DevopsSecretService {
     private V1Secret initV1Secret(DevopsSecretE devopsSecretE) {
         V1Secret secret = new V1Secret();
         secret.setApiVersion("v1");
-        secret.setKind("Secret");
+        secret.setKind(SECRET);
         V1ObjectMeta metadata = new V1ObjectMeta();
         metadata.setName(devopsSecretE.getName());
         secret.setMetadata(metadata);
@@ -163,7 +161,7 @@ public class DevopsSecretServiceImpl implements DevopsSecretService {
         ObjectOperation<V1Secret> objectOperation = new ObjectOperation<>();
         objectOperation.setType(v1Secret);
         objectOperation.operationEnvGitlabFile("sct-" + devopsSecretE.getName(), gitlabEnvGroupProjectId,
-                isCreate ? "create" : "update", userAttrE.getGitlabUserId(), devopsSecretE.getId(), SECRET,
+                isCreate ? CREATE : UPDATE, userAttrE.getGitlabUserId(), devopsSecretE.getId(), SECRET,
                 devopsSecretE.getEnvId(), path);
 
         DevopsSecretE afterDevopsSecretE = devopsSecretRepository
@@ -214,6 +212,14 @@ public class DevopsSecretServiceImpl implements DevopsSecretService {
                 .queryByEnvIdAndResource(devopsEnvironmentE.getId(), secretId, SECRET);
         if (devopsEnvFileResourceE == null) {
             devopsSecretRepository.deleteSecret(secretId);
+            if (gitlabRepository.getFile(TypeUtil.objToInteger(devopsEnvironmentE.getGitlabEnvProjectId()), "master",
+                    "sct-" + devopsSecretE.getName() + ".yaml")) {
+                gitlabRepository.deleteFile(
+                        TypeUtil.objToInteger(devopsEnvironmentE.getGitlabEnvProjectId()),
+                        "sct-" + devopsSecretE.getName() + ".yaml",
+                        "DELETE FILE",
+                        TypeUtil.objToInteger(userAttrE.getGitlabUserId()));
+            }
             return true;
         }
         List<DevopsEnvFileResourceE> devopsEnvFileResourceES = devopsEnvFileResourceRepository
@@ -221,9 +227,12 @@ public class DevopsSecretServiceImpl implements DevopsSecretService {
 
         // 如果对象所在文件只有一个对象，则直接删除文件,否则把对象从文件中去掉，更新文件
         if (devopsEnvFileResourceES.size() == 1) {
-            gitlabRepository.deleteFile(TypeUtil.objToInteger(devopsEnvironmentE.getGitlabEnvProjectId()),
-                    devopsEnvFileResourceE.getFilePath(), "DELETE FILE",
-                    TypeUtil.objToInteger(userAttrE.getGitlabUserId()));
+            if (gitlabRepository.getFile(TypeUtil.objToInteger(devopsEnvironmentE.getGitlabEnvProjectId()), "master",
+                    devopsEnvFileResourceE.getFilePath())) {
+                gitlabRepository.deleteFile(TypeUtil.objToInteger(devopsEnvironmentE.getGitlabEnvProjectId()),
+                        devopsEnvFileResourceE.getFilePath(), "DELETE FILE",
+                        TypeUtil.objToInteger(userAttrE.getGitlabUserId()));
+            }
         } else {
             ObjectOperation<V1Secret> objectOperation = new ObjectOperation<>();
             V1Secret v1Secret = new V1Secret();
@@ -257,13 +266,11 @@ public class DevopsSecretServiceImpl implements DevopsSecretService {
 
         // 处理secret对象
         DevopsSecretE devopsSecretE = handleSecret(secretReqDTO);
-        // 初始化V1Secret对象
-        V1Secret v1Secret = initV1Secret(devopsSecretE);
 
         // 创建secret
         Long secretId = devopsSecretRepository.create(devopsSecretE).getId();
         DevopsEnvCommandE devopsEnvCommandE = new DevopsEnvCommandE();
-        devopsEnvCommandE.setCommandType("create");
+        devopsEnvCommandE.setCommandType(CREATE);
         devopsEnvCommandE.setStatus(devopsSecretE.getStatus());
         devopsEnvCommandE.setObjectId(secretId);
         devopsEnvCommandE.setObject("secret");
@@ -289,8 +296,6 @@ public class DevopsSecretServiceImpl implements DevopsSecretService {
 
         // 处理secret对象
         DevopsSecretE devopsSecretE = handleSecret(secretReqDTO);
-        // 初始化V1Secret对象
-        V1Secret v1Secret = initV1Secret(devopsSecretE);
 
         DevopsEnvCommandE devopsEnvCommandE = initDevopsEnvCommandE(UPDATE);
         // 更新secret对象
